Revert "Dynamic entitlement agent (#116125)"
This reverts commit 338c0538b7ba986f3bace7e3d0da39d636aa0b15.
