+++ build-tools/src/main/java/org/elasticsearch/gradle/testclusters/RunTask.java
@@ -42,7 +42,6 @@ public abstract class RunTask extends DefaultTestClustersTask {
 
     private Boolean debug = false;
     private Boolean cliDebug = false;
-    private Boolean entitlementsEnabled = false;
     private Boolean apmServerEnabled = false;
 
     private Boolean preserveData = false;
@@ -70,14 +69,6 @@ public abstract class RunTask extends DefaultTestClustersTask {
         this.cliDebug = enabled;
     }
 
-    @Option(
-        option = "entitlements",
-        description = "Use the Entitlements agent system in place of SecurityManager to enforce sandbox policies."
-    )
-    public void setEntitlementsEnabled(boolean enabled) {
-        this.entitlementsEnabled = enabled;
-    }
-
     @Input
     public Boolean getDebug() {
         return debug;
@@ -86,11 +77,6 @@ public abstract class RunTask extends DefaultTestClustersTask {
     @Input
     public Boolean getCliDebug() {
         return cliDebug;
-    }
-
-    @Input
-    public Boolean getEntitlementsEnabled() {
-        return entitlementsEnabled;
     }
 
     @Input
@@ -239,9 +225,6 @@ public abstract class RunTask extends DefaultTestClustersTask {
         }
         if (cliDebug) {
             enableCliDebug();
-        }
-        if (entitlementsEnabled) {
-            enableEntitlements();
         }
     }
