+++ build-tools/src/main/java/org/elasticsearch/gradle/testclusters/TestClustersAware.java
@@ -74,12 +74,4 @@ public interface TestClustersAware extends Task {
             }
         }
     }
-
-    default void enableEntitlements() {
-        for (ElasticsearchCluster cluster : getClusters()) {
-            for (ElasticsearchNode node : cluster.getNodes()) {
-                node.cliJvmArgs("-Des.entitlements.enabled=true");
-            }
-        }
-    }
 }
