+++ distribution/build.gradle
@@ -262,7 +262,7 @@ configure(subprojects.findAll { ['archives', 'packages'].contains(it.name) }) {
    *             Properties to expand when copying packaging files             *
    *****************************************************************************/
   configurations {
-    ['libs', 'libsVersionChecker', 'libsCliLauncher', 'libsServerCli', 'libsWindowsServiceCli', 'libsPluginCli', 'libsKeystoreCli', 'libsSecurityCli', 'libsGeoIpCli', 'libsAnsiConsole', 'libsNative', 'libsEntitlementAgent', 'libsEntitlementBridge'].each {
+    ['libs', 'libsVersionChecker', 'libsCliLauncher', 'libsServerCli', 'libsWindowsServiceCli', 'libsPluginCli', 'libsKeystoreCli', 'libsSecurityCli', 'libsGeoIpCli', 'libsAnsiConsole', 'libsNative'].each {
       create(it) {
         canBeConsumed = false
         canBeResolved = true
@@ -292,8 +292,6 @@ configure(subprojects.findAll { ['archives', 'packages'].contains(it.name) }) {
     libsSecurityCli project(':x-pack:plugin:security:cli')
     libsGeoIpCli project(':distribution:tools:geoip-cli')
     libsNative project(':libs:native:native-libraries')
-    libsEntitlementAgent project(':libs:entitlement:agent')
-    libsEntitlementBridge project(':libs:entitlement:bridge')
   }
 
   project.ext {
@@ -337,12 +335,6 @@ configure(subprojects.findAll { ['archives', 'packages'].contains(it.name) }) {
           if (os != null) {
             include (os + '-' + architecture + '/*')
           }
-        }
-        into('entitlement-agent') {
-          from(configurations.libsEntitlementAgent)
-        }
-        into('entitlement-bridge') {
-          from(configurations.libsEntitlementBridge)
         }
       }
     }
