+++ distribution/tools/server-cli/src/main/java/org/elasticsearch/server/cli/SystemJvmOptions.java
@@ -12,11 +12,9 @@ package org.elasticsearch.server.cli;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.util.concurrent.EsExecutors;
 
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 final class SystemJvmOptions {
@@ -24,8 +22,8 @@ final class SystemJvmOptions {
     static List<String> systemJvmOptions(Settings nodeSettings, final Map<String, String> sysprops) {
         String distroType = sysprops.get("es.distribution.type");
         boolean isHotspot = sysprops.getOrDefault("sun.management.compiler", "").contains("HotSpot");
-        boolean useEntitlements = Boolean.parseBoolean(sysprops.getOrDefault("es.entitlements.enabled", "false"));
-        return Stream.of(
+
+        return Stream.concat(
             Stream.of(
                 /*
                  * Cache ttl in seconds for positive DNS lookups noting that this overrides the JDK security property
@@ -37,6 +35,8 @@ final class SystemJvmOptions {
                  * networkaddress.cache.negative ttl; set to -1 to cache forever.
                  */
                 "-Des.networkaddress.cache.negative.ttl=10",
+                // Allow to set the security manager.
+                "-Djava.security.manager=allow",
                 // pre-touch JVM emory pages during initialization
                 "-XX:+AlwaysPreTouch",
                 // explicitly set the stack size
@@ -61,17 +61,15 @@ final class SystemJvmOptions {
                 "-Dlog4j2.disable.jmx=true",
                 "-Dlog4j2.formatMsgNoLookups=true",
                 "-Djava.locale.providers=CLDR",
+                maybeEnableNativeAccess(),
+                maybeOverrideDockerCgroup(distroType),
+                maybeSetActiveProcessorCount(nodeSettings),
+                setReplayFile(distroType, isHotspot),
                 // Pass through distribution type
                 "-Des.distribution.type=" + distroType
             ),
-            maybeEnableNativeAccess(),
-            maybeOverrideDockerCgroup(distroType),
-            maybeSetActiveProcessorCount(nodeSettings),
-            maybeSetReplayFile(distroType, isHotspot),
-            maybeWorkaroundG1Bug(),
-            maybeAllowSecurityManager(),
-            maybeAttachEntitlementAgent(useEntitlements)
-        ).flatMap(s -> s).toList();
+            maybeWorkaroundG1Bug()
+        ).filter(e -> e.isEmpty() == false).collect(Collectors.toList());
     }
 
     /*
@@ -88,42 +86,42 @@ final class SystemJvmOptions {
      * that cgroup statistics are available for the container this process
      * will run in.
      */
-    private static Stream<String> maybeOverrideDockerCgroup(String distroType) {
+    private static String maybeOverrideDockerCgroup(String distroType) {
         if ("docker".equals(distroType)) {
-            return Stream.of("-Des.cgroups.hierarchy.override=/");
+            return "-Des.cgroups.hierarchy.override=/";
         }
-        return Stream.empty();
+        return "";
     }
 
-    private static Stream<String> maybeSetReplayFile(String distroType, boolean isHotspot) {
+    private static String setReplayFile(String distroType, boolean isHotspot) {
         if (isHotspot == false) {
             // the replay file option is only guaranteed for hotspot vms
-            return Stream.empty();
+            return "";
         }
         String replayDir = "logs";
         if ("rpm".equals(distroType) || "deb".equals(distroType)) {
             replayDir = "/var/log/elasticsearch";
         }
-        return Stream.of("-XX:ReplayDataFile=" + replayDir + "/replay_pid%p.log");
+        return "-XX:ReplayDataFile=" + replayDir + "/replay_pid%p.log";
     }
 
     /*
      * node.processors determines thread pool sizes for Elasticsearch. When it
      * is set, we need to also tell the JVM to respect a different value
      */
-    private static Stream<String> maybeSetActiveProcessorCount(Settings nodeSettings) {
+    private static String maybeSetActiveProcessorCount(Settings nodeSettings) {
         if (EsExecutors.NODE_PROCESSORS_SETTING.exists(nodeSettings)) {
             int allocated = EsExecutors.allocatedProcessors(nodeSettings);
-            return Stream.of("-XX:ActiveProcessorCount=" + allocated);
+            return "-XX:ActiveProcessorCount=" + allocated;
         }
-        return Stream.empty();
+        return "";
     }
 
-    private static Stream<String> maybeEnableNativeAccess() {
+    private static String maybeEnableNativeAccess() {
         if (Runtime.version().feature() >= 21) {
-            return Stream.of("--enable-native-access=org.elasticsearch.nativeaccess,org.apache.lucene.core");
+            return "--enable-native-access=org.elasticsearch.nativeaccess,org.apache.lucene.core";
         }
-        return Stream.empty();
+        return "";
     }
 
     /*
@@ -135,38 +133,5 @@ final class SystemJvmOptions {
             return Stream.of("-XX:+UnlockDiagnosticVMOptions", "-XX:G1NumCollectionsKeepPinned=10000000");
         }
         return Stream.of();
-    }
-
-    private static Stream<String> maybeAllowSecurityManager() {
-        // Will become conditional on useEntitlements once entitlements can run without SM
-        return Stream.of("-Djava.security.manager=allow");
-    }
-
-    private static Stream<String> maybeAttachEntitlementAgent(boolean useEntitlements) {
-        if (useEntitlements == false) {
-            return Stream.empty();
-        }
-
-        Path dir = Path.of("lib", "entitlement-bridge");
-        if (Files.exists(dir) == false) {
-            throw new IllegalStateException("Directory for entitlement bridge jar does not exist: " + dir);
-        }
-        String bridgeJar;
-        try (var s = Files.list(dir)) {
-            var candidates = s.limit(2).toList();
-            if (candidates.size() != 1) {
-                throw new IllegalStateException("Expected one jar in " + dir + "; found " + candidates.size());
-            }
-            bridgeJar = candidates.get(0).toString();
-        } catch (IOException e) {
-            throw new IllegalStateException("Failed to list entitlement jars in: " + dir, e);
-        }
-        return Stream.of(
-            "-Des.entitlements.enabled=true",
-            "-XX:+EnableDynamicAgentLoading",
-            "-Djdk.attach.allowAttachSelf=true",
-            "--patch-module=java.base=" + bridgeJar,
-            "--add-exports=java.base/org.elasticsearch.entitlement.bridge=org.elasticsearch.entitlement"
-        );
     }
 }
