+++ libs/entitlement/agent/build.gradle
@@ -6,18 +6,51 @@
  * your election, the "Elastic License 2.0", the "GNU Affero General Public
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
+
+import static java.util.stream.Collectors.joining
+
 apply plugin: 'elasticsearch.build'
+apply plugin: 'elasticsearch.embedded-providers'
+
+embeddedProviders {
+  impl 'entitlement-agent', project(':libs:entitlement:agent:impl')
+}
+
+configurations {
+  entitlementBridge
+}
 
 dependencies {
+  entitlementBridge project(":libs:entitlement:bridge")
   compileOnly project(":libs:core")
   compileOnly project(":libs:entitlement")
-  compileOnly project(":libs:entitlement:bridge")
+  testImplementation project(":test:framework")
+  testImplementation project(":libs:entitlement:bridge")
+  testImplementation project(":libs:entitlement:agent:impl")
+}
+
+tasks.named('test').configure {
+  systemProperty "tests.security.manager", "false"
+  dependsOn('jar')
+
+  // Register an argument provider to avoid eager resolution of configurations
+  jvmArgumentProviders.add(new CommandLineArgumentProvider() {
+    @Override
+    Iterable<String> asArguments() {
+      return ["-javaagent:${tasks.jar.archiveFile.get()}", "-Des.entitlements.bridgeJar=${configurations.entitlementBridge.singleFile}"]
+    }
+  })
+
+
+  // The Elasticsearch build plugin automatically adds all compileOnly deps as testImplementation.
+  // We must not add the bridge this way because it is also on the boot classpath, and that would lead to jar hell.
+  classpath -= files(configurations.entitlementBridge)
 }
 
 tasks.named('jar').configure {
   manifest {
     attributes(
-      'Agent-Class': 'org.elasticsearch.entitlement.agent.EntitlementAgent'
+      'Premain-Class': 'org.elasticsearch.entitlement.agent.EntitlementAgent'
       , 'Can-Retransform-Classes': 'true'
     )
   }
