+++ libs/entitlement/agent/impl/build.gradle
@@ -7,14 +7,21 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-module org.elasticsearch.entitlement.runtime {
-    requires org.elasticsearch.entitlement.bridge;
-    requires org.elasticsearch.xcontent;
-    requires org.elasticsearch.server;
+apply plugin: 'elasticsearch.build'
 
-    exports org.elasticsearch.entitlement.runtime.api;
-
-    provides org.elasticsearch.entitlement.api.EntitlementChecks
-        with
-            org.elasticsearch.entitlement.runtime.api.ElasticsearchEntitlementManager;
+dependencies {
+  compileOnly project(':libs:entitlement:agent')
+  implementation 'org.ow2.asm:asm:9.7'
+  testImplementation project(":test:framework")
+  testImplementation project(":libs:entitlement:bridge")
+  testImplementation 'org.ow2.asm:asm-util:9.7'
 }
+
+tasks.named('test').configure {
+  systemProperty "tests.security.manager", "false"
+}
+
+tasks.named('forbiddenApisMain').configure {
+  replaceSignatureFiles 'jdk-signatures'
+}
+
