+++ libs/entitlement/agent/src/main/java/org/elasticsearch/entitlement/instrumentation/MethodKey.java
