--- libs/entitlement/asm-provider/README.md
@@ -1,2 +1,0 @@
-This module uses the ASM library to implement various things, including bytecode instrumentation.
-It is loaded using the Embedded Provider Gradle plugin.
