+++ libs/entitlement/bridge/build.gradle
@@ -9,6 +9,8 @@
 
 apply plugin: 'elasticsearch.build'
 
+dependencies {
+}
 
 tasks.named('forbiddenApisMain').configure {
   replaceSignatureFiles 'jdk-signatures'
