+++ libs/entitlement/build.gradle
@@ -9,20 +9,12 @@
 apply plugin: 'elasticsearch.build'
 apply plugin: 'elasticsearch.publish'
 
-apply plugin: 'elasticsearch.embedded-providers'
-
-embeddedProviders {
-  impl 'entitlement', project(':libs:entitlement:asm-provider')
-}
-
 dependencies {
   compileOnly project(':libs:core') // For @SuppressForbidden
-  compileOnly project(':libs:logging')
   compileOnly project(":libs:x-content") // for parsing policy files
+  compileOnly project(':server') // To access the main server module for special permission checks
   compileOnly project(':libs:entitlement:bridge')
-  testImplementation(project(":test:framework")) {
-    exclude group: 'org.elasticsearch', module: 'entitlement'
-  }
+  testImplementation project(":test:framework")
 }
 
 tasks.named('forbiddenApisMain').configure {
