+++ server/src/main/java/module-info.java
@@ -31,7 +31,6 @@ module org.elasticsearch.server {
     requires org.elasticsearch.grok;
     requires org.elasticsearch.tdigest;
     requires org.elasticsearch.simdvec;
-    requires org.elasticsearch.entitlement;
 
     requires hppc;
     requires HdrHistogram;
