+++ server/src/main/java/org/elasticsearch/bootstrap/Elasticsearch.java
@@ -30,7 +30,6 @@ import org.elasticsearch.common.util.concurrent.RunOnce;
 import org.elasticsearch.core.AbstractRefCounted;
 import org.elasticsearch.core.IOUtils;
 import org.elasticsearch.core.SuppressForbidden;
-import org.elasticsearch.entitlement.bootstrap.EntitlementBootstrap;
 import org.elasticsearch.env.Environment;
 import org.elasticsearch.index.IndexVersion;
 import org.elasticsearch.jdk.JarHell;
@@ -199,16 +198,12 @@ class Elasticsearch {
             VectorUtil.class
         );
 
-        if (Boolean.parseBoolean(System.getProperty("es.entitlements.enabled"))) {
-            EntitlementBootstrap.bootstrap();
-        } else {
-            // install SM after natives, shutdown hooks, etc.
-            org.elasticsearch.bootstrap.Security.configure(
-                nodeEnv,
-                SECURITY_FILTER_BAD_DEFAULTS_SETTING.get(args.nodeSettings()),
-                args.pidFile()
-            );
-        }
+        // install SM after natives, shutdown hooks, etc.
+        org.elasticsearch.bootstrap.Security.configure(
+            nodeEnv,
+            SECURITY_FILTER_BAD_DEFAULTS_SETTING.get(args.nodeSettings()),
+            args.pidFile()
+        );
     }
 
     private static void ensureInitialized(Class<?>... classes) {
